requirements.txt
flask
requests
beautifulsoup4
pycryptodome
numpy
opencv-python-headless
python-dotenv
google-cloud-storage
gunicorn
google-cloud-storage